% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset-world.R
\name{inset_world}
\alias{inset_world}
\title{Create a world map emphasizing one country.}
\usage{
inset_world(emphasize, ...)
}
\arguments{
\item{emphasize}{Character vector with country names as found in the "world"
database in the `maps` package.}
}
\value{
A ggplot object.
}
\description{
This function creates a `ggplot2` plot showing a map of the world on the
right panel and an inset showing a country or countries on the left panel.
}
\details{
The maps data are obtained from the `maps` package.
}
\examples{
\dontrun{
# Load the package
library(package = "insetmap")

# Prepare the map
inset1 <- inset_world(emphasize = c("Guatemala", "Costa Rica"))
inset1

# Modify a map

# Load ggplot2 package
library(package = "ggplot2")

# Modify the produced plot
inset2 <- inset1 +
  scale_fill_manual(values = c("#ffaaaa", "blue"))
inset2
}
}
