% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insetmap.R
\docType{package}
\name{insetmap-package}
\alias{insetmap}
\alias{insetmap-package}
\title{insetmap: Easily create simple inset maps}
\description{
Prepare inset maps using simple wrappers based on existing tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odeleongt/insetmap}
  \item Report bugs at \url{https://github.com/odeleongt/insetmap/issues/}
}

}
\author{
\strong{Maintainer}: Oscar de León \email{odeleon@ces.uvg.edu.gt}

}
\keyword{internal}
