% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{read_sf}
\alias{reexports}
\alias{st_transform}
\alias{reexports}
\alias{st_set_crs}
\alias{reexports}
\alias{mutate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{sf}{\code{\link[sf]{read_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_set_crs}}}
}}

